// SGUEES-SPA/src/app/pages/General/gen-municipio/models/gen-municipio.ts
export interface GenMunicipio {
	CORR_EMPRESA: number;
	CORR_PAIS: number;
	CORR_DEPTO: number;
	CORR_MUNICIPIO: number;
	NOMBRE_MUNICIPIO: string;
	CODIGO_MUNICIPIO: string;
	USUARIO_CREA: string;
	ESTACION_CREA: string;
	FECHA_CREA: Date;
	USUARIO_ACTU: string;
	ESTACION_ACTU: string;
	FECHA_ACTU: Date;
	NOMBRE_DEPTO: string;
	NOMBRE_PAIS: string;
}